+++ packages/dendron-11ty/bin/build-styles.js
@@ -1,4 +1,4 @@
-const sass = require('node-sass');
+const sass = require('sass');
 const fs = require('fs-extra');
 const util = require('util');
 const {getSiteOutputPath} = require("../libs/utils");
@@ -6,11 +6,10 @@ const path = require("path");
 
 
 const buildSass = async function () {
-  const renderSass = util.promisify(sass.render);
   const inputFile = path.join(getSiteOutputPath(), "raw-assets", "sass", "just-the-docs-default.scss")
   const outputFile = path.join(getSiteOutputPath(), "assets", "css", "just-the-docs-default.css")
   const isProduction = process.env.ELEVENTY_ENV;
-  const { css } = await renderSass({
+  const {css} = sass.renderSync({
     file: inputFile,
     includePaths: [
       'node_modules/foundation-sites/scss/',
@@ -18,9 +17,8 @@ const buildSass = async function () {
       'node_modules/hamburgers/_sass/hamburgers',
       'node_modules/prismjs/themes',
     ],
-    outputStyle: isProduction ? 'compressed' : 'nested',
-  });
-
+    outputStyle: isProduction ? 'compressed' : 'expanded',
+  })
   // This is a hint to know if this is a first run, in which case
   // we don't need to tell browserSync to update.
   const fileExisted = await fs.pathExists(outputFile);