+++ packages/plugin-core/scripts/package.sh
@@ -3,6 +3,9 @@
 git reset --hard
 git clean -f
 git pull
+
+# don't install in workspace
+rm ../../package.json
 # conform to vscode naming  convention
 sed  -ibak 's/@dendronhq.plugin-core/dendron/' package.json
 sed  -ibak 's/out\/extension/dist\/extension/' package.json
@@ -11,5 +14,5 @@ cat package.json | jq '.repository = { "url": "https://github.com/dendronhq/dend
 mv tmp.json package.json
 
 #./scripts/sync_static.sh
-npm install
-vsce package
+yarn install --no-lockfile
+vsce package --yarn