+++ packages/plugin-core/src/commands/SiteBuild.ts
@@ -1,8 +1,7 @@
 import { BuildSiteV2CLICommandOpts } from "@dendronhq/dendron-cli";
-import { execa } from "@dendronhq/engine-server";
 import { env, Uri, window } from "vscode";
 import { DENDRON_COMMANDS } from "../constants";
-import { checkPreReq, getSiteRootDirPath } from "../utils/site";
+import { buildSite, checkPreReq, getSiteRootDirPath } from "../utils/site";
 import { DendronWorkspace } from "../workspace";
 import { BasicCommand } from "./base";
 
@@ -19,18 +18,18 @@ export class SiteBuildCommand extends BasicCommand<CommandOpts, CommandOutput> {
 
   async sanityCheck() {
     return checkPreReq();
+    // return undefined;
   }
 
   async execute(_opts?: CommandOpts) {
     const wsRoot = DendronWorkspace.wsRoot();
     window.showInformationMessage("building...");
     const port = DendronWorkspace.instance().port!;
-    // TODO: show progress
-    const cmdBuild = `dendron-cli buildSiteV2 --wsRoot ${wsRoot} --stage prod --enginePort ${port}`.split(
-      " "
-    );
-    await execa("npx", cmdBuild, {
-      cwd: wsRoot,
+    await buildSite({
+      wsRoot,
+      stage: "prod",
+      enginePort: port,
+      serve: false,
     });
   }
   async showResponse() {