+++ packages/plugin-core/webpack-require-hack.js
@@ -1,0 +1,5 @@
+const importEleventy = (importPath) => {
+    const eleventy = require(importPath);
+    return eleventy;
+};
+module.exports = importEleventy;