+++ packages/plugin-core/webpack.config.js
@@ -22,34 +22,28 @@ const config = {
     __dirname: false,
   },
   devtool: "source-map",
-  externals: {
-    vscode: "commonjs vscode", // the vscode-module is created on-the-fly and must be excluded. Add other modules that cannot be webpack'ed, 📖 -> https://webpack.js.org/configuration/externals/
-    "pino-pretty": "pino-pretty",
-  },
+  externals: [
+    {
+      vscode: "commonjs vscode", // the vscode-module is created on-the-fly and must be excluded. Add other modules that cannot be webpack'ed, 📖 -> https://webpack.js.org/configuration/externals/
+      "pino-pretty": "pino-pretty",
+    },
+    /(@dendronhq|packages)\/dendron-11ty$/,
+    /\.\/webpack-require-hack/
+  ],
   resolve: {
     // support reading TypeScript and JavaScript files, 📖 -> https://github.com/TypeStrong/ts-loader
     extensions: [".ts", ".js"],
   },
   plugins: [
-    new CopyPlugin({
-      patterns: [
-        {
-          from: path.join("node_modules", "@dendronhq", "lsp-server"),
-          to: "lsp-server",
-        },
-      ],
-    }),
-    new CopyPlugin({
-      patterns: [{ from: path.join("assets", "static"), to: "static" }],
-    }),
+    // new CopyPlugin({
+    //   patterns: [{ from: path.join("assets", "static"), to: "static" }],
+    // }),
     new CopyPlugin({
       patterns: [{ from: path.join("assets", "dendron-ws"), to: "dendron-ws" }],
     }),
-    new IgnorePlugin({
-      // resourceRegExp: /^\.\/locale$/,
-      resourceRegExp: /^@dendronhq\/dendron-11ty$/,
-      contextRegExp: /commands$/
-    })
+    new CopyPlugin({
+      patterns: [{ from: "webpack-require-hack.js", to: "webpack-require-hack.js" }],
+    }),
   ],
   module: {
     rules: [